chore: rework go dependencies (#321)
Reworked the Go dependencies for this repository so that we do not need
to list the transitive dependencies in `MODULE.bazel`. Because
`github.com/ekalinin/github-markdown-toc.go` references its dependencies
in its `main` package, `go get ...` does not add the transitive
dependencies to `go.mod`. The bzlmod-aware `update-repos` does not find
the transitive dependencies. To fix this, we add direct dependency for
`github.com/ekalinin/github-markdown-toc.go` to this repo and use it in
`cmd/go_deps`.

Related to https://github.com/bazelbuild/bazel-gazelle/issues/1585.